move revision checker to helpers
